requirements.txt
python-multipart
fastapi
uvicorn
spacy
langdetect
openai
python-dotenv
sqlalchemy
alembic
psycopg2-binary
pydantic
